Change how modules are created.
Now, a module object is passed in for the first parameter of the
module construction function. That is the object one should append
the api to.

Use the module.dependency function to declare a runtime dependency
(as opposed to the "compile" time dependencies that are declared in
the second parameter to binfo._render);